refactor(dev): migrate from LocalStack to GoldenStack
Replace localstack/localstack with the GoldenStack drop-in replacement
(ECR image 764525110978.dkr.ecr.us-west-2.amazonaws.com/goldenstack:1).

All prior LocalStack env vars (LOCALSTACK, LOCALSTACK_URL), option names
(useLocalstack), and npm scripts (init-localstack, localstack) are kept
as aliases so existing consumers have no breaking change.
