+++ docker-compose.yml
@@ -1,16 +1,14 @@
 version: "3.8"
 
 services:
-  localstack:
-    container_name: "${LOCALSTACK_DOCKER_NAME-localstack_main}"
-    image: localstack/localstack:1.4.0
+  goldenstack:
+    container_name: goldenstack
+    image: 764525110978.dkr.ecr.us-west-2.amazonaws.com/goldenstack:1
     ports:
-      - "4566:4566"            # LocalStack Gateway
-      - "9999:8080"
+      - "4566:4566"
     environment:
       - DEBUG=${DEBUG-}
       - SERVICES=dynamodb,s3
       - DOCKER_HOST=unix:///var/run/docker.sock
     volumes:
-      - "${TMPDIR:-/tmp/localstack}:/tmp/localstack"
       - "/var/run/docker.sock:/var/run/docker.sock"
