+++ lib/config.js
@@ -33,9 +33,9 @@ module.exports = fp(
       config.acl = DEFAULT_ACL;
     }
 
-    const localstack = process.env.LOCALSTACK || 0;
-    if (localstack || opts.useLocalstack) {
-      const endpoint = process.env.LOCALSTACK_URL || 'http://localhost:4566';
+    const goldenstack = process.env.GOLDENSTACK || process.env.LOCALSTACK || 0;
+    if (goldenstack || opts.useGoldenstack || opts.useLocalstack) {
+      const endpoint = process.env.GOLDENSTACK_URL || process.env.LOCALSTACK_URL || 'http://localhost:4566';
       const region = process.env.AWS_REGION || 'us-west-2';
       const credentials = { accessKeyId: 'foo', secretAccessKey: 'bar' };
       config.dynamo = { endpoint, region, credentials };
